chore: move file-storage tests
